refactor(security): Remove local and use shared DTO for 'auth' package
