--- mp-security/src/main/java/ru/pavbatol/myplace/auth/dto/AuthDtoResponse.java
@@ -1,10 +1,0 @@
-package ru.pavbatol.myplace.auth.dto;
-
-import lombok.Value;
-
-@Value
-public class AuthDtoResponse {
-    String type = "Bearer";
-    String accessToken;
-    String refreshToken;
-}
