+++ mp-security/src/main/java/ru/pavbatol/myplace/auth/service/AuthServiceImpl.java
@@ -18,12 +18,12 @@ import org.springframework.stereotype.Service;
 import ru.pavbatol.myplace.app.exception.BadRequestException;
 import ru.pavbatol.myplace.app.exception.NotFoundException;
 import ru.pavbatol.myplace.app.exception.RedisException;
-import ru.pavbatol.myplace.auth.dto.AuthDtoRequest;
-import ru.pavbatol.myplace.auth.dto.AuthDtoResponse;
 import ru.pavbatol.myplace.auth.model.AccessTokenDetails;
 import ru.pavbatol.myplace.auth.repository.AccessTokenRedisRepository;
 import ru.pavbatol.myplace.auth.repository.RefreshTokenRedisRepository;
 import ru.pavbatol.myplace.jwt.JwtProvider;
+import ru.pavbatol.myplace.shared.dto.security.auth.AuthDtoRequest;
+import ru.pavbatol.myplace.shared.dto.security.auth.AuthDtoResponse;
 import ru.pavbatol.myplace.user.model.User;
 import ru.pavbatol.myplace.user.model.UserAuthenticationPrincipal;
 import ru.pavbatol.myplace.user.repository.UserJpaRepository;
